[core] Move GLContextMode to its own Renderer header
