+++ include/mbgl/map/mode.hpp
@@ -14,15 +14,6 @@ enum class MapMode : EnumType {
     Still, // a once-off still image
 };
 
-// We can avoid redundant GL calls when it is known that the GL context is not
-// being shared. In a shared GL context case, we need to make sure that the
-// correct GL configurations are in use - they might have changed between render
-// calls.
-enum class GLContextMode : EnumType {
-    Unique,
-    Shared,
-};
-
 // We can choose to constrain the map both horizontally or vertically, or only
 // vertically e.g. while panning.
 enum class ConstrainMode : EnumType {
